add in pieces from client guardrails PR

+++ packages/evidence/cli.js
@@ -15,9 +15,12 @@ const populateTemplate = function () {
 
 	fs.ensureDirSync('./.evidence/template/');
 
-	// empty the template directory, except any local settings, or telemetry profile that already exist.
+	// empty the template directory, except:
+	// - local settings
+	// - telemetry profile
+	// - static folder (mainly to preserve the data directory)
 	fs.readdirSync('./.evidence/template/').forEach((file) => {
-		if (file != 'evidence.settings.json' && file != '.profile.json')
+		if (file != 'evidence.settings.json' && file != '.profile.json' && file != 'static')
 			fs.removeSync(path.join('./.evidence/template/', file));
 	});
